NoticeRow 구현
